third_party/WebKit/LayoutTests/fast/canvas/linearGradient-infinite-values-expected.txt

This test checks createLinearGradient with infinite values

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.createLinearGradient(0, 0, 100, NaN) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y1 is not a number..
PASS ctx.createLinearGradient(0, 0, 100, Infinity) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y1 is infinite..
PASS ctx.createLinearGradient(0, 0, 100, -Infinity) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y1 is infinite..
PASS ctx.createLinearGradient(0, 0, NaN, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x1 is not a number..
PASS ctx.createLinearGradient(0, 0, Infinity, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x1 is infinite..
PASS ctx.createLinearGradient(0, 0, -Infinity, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x1 is infinite..
PASS ctx.createLinearGradient(0, NaN, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y0 is not a number..
PASS ctx.createLinearGradient(0, Infinity, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y0 is infinite..
PASS ctx.createLinearGradient(0, -Infinity, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The y0 is infinite..
PASS ctx.createLinearGradient(NaN, 0, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x0 is not a number..
PASS ctx.createLinearGradient(Infinity, 0, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x0 is infinite..
PASS ctx.createLinearGradient(-Infinity, 0, 100, 100) threw exception NotSupportedError: Failed to execute 'createLinearGradient' on 'CanvasRenderingContext2D': The x0 is infinite..
PASS successfullyParsed is true

TEST COMPLETE
